+++ docs/README.md
@@ -28,7 +28,7 @@
     - 딜러의 이름은 딜러다. 
     - [ ] 카드 리스트 
       - 각 사용자의 전체 소유 카드를 가진다.
-  - [ ] 카드 분배기
+  - [x] 카드 분배기
     - 한 사람 당 한 번에 하나의 카드를 배분해준다.
     - [x] 카드 덱
       - 52장의 초기 개수를 가진다.
