+++ src/test/java/util/CardDeckMakerTest.java
@@ -1,17 +1,21 @@
-package domain;
+package util;
 
+import domain.Card;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class CardDeckTest {
+public class CardDeckMakerTest {
 
     @Test
     @DisplayName("Card 덱은 52 장의 초기 개수를 가진다.")
     void createCardDeckSuccess() {
-        CardDeck cardDeck = new CardDeck();
+        List<Card> cards = CardDeckMaker.generate();
 
-        assertThat(cardDeck.getSize()).isEqualTo(52);
+        assertThat(cards.size()).isEqualTo(52);
     }
+
 }
